+++ NEWS.md
@@ -1,7 +1,9 @@
 # datamods 1.3.0
 
-* Added module to read flat data from URLs `import_url_*()`.
+* New module to read flat data from URLs `import_url_*()`.
 * Error messages displayed to the user are more informative on the actual error.
+* `filter_data_server()`: new argument `value_na` to set default value for NA's filters widgets.
+
 
 # datamods 1.2.0