+++ NEWS.md
@@ -1,10 +1,11 @@
-# datamods 1.3.0
+# datamods (unreleased)
 
 * New module to read flat data from URLs `import_url_*()`.
 * Error messages displayed to the user are more informative on the actual error.
 * `filter_data_server()`: new argument `value_na` to set default value for NA's filters widgets.
 * i18n: chinese translations added, thanks to [@xmusphlkg](https://github.com/xmusphlkg).
 * i18n: spanish translations added, thanks to [@drdsdaniel](https://github.com/drdsdaniel).
+* i18n: german translations added, thanks to [@SteEcker](https://github.com/SteEcker) and [@joerghenkebuero](https://github.com/joerghenkebuero).
 
 
 # datamods 1.2.0